Started on main class and changed heightOfPlan from int to double

+++ src/Greenest_Assignment_1/Plant_Super_Sub/Cactus.java
@@ -6,7 +6,7 @@ public class Cactus extends Plant {
 
     private final Liquid_Type liquid_Type = Liquid_Type.MINERALWATER;
 
-    public Cactus(String plantType, String plantName, int heightOfPlant) {
+    public Cactus(String plantType, String plantName, double heightOfPlant) {
         super(plantType, plantName, heightOfPlant);
     }
